+++ aa-flow/web-test/src/config/networks.ts
@@ -47,6 +47,7 @@ export const NETWORKS: Record<string, NetworkConfig> = {
       entryPoint: '0x5FF137D4b0FDCD49DcA30c7CF57E578a026d2789',
       factory: '0x9406Cc6185a346906296840746125a0E44976454',
     },
+    bundlerUrl: import.meta.env.VITE_BUNDLER_URL || 'https://rundler-superrelay.fly.dev',
   },
   opMainnet: {
     id: 10,
@@ -62,6 +63,7 @@ export const NETWORKS: Record<string, NetworkConfig> = {
       entryPoint: '0x5FF137D4b0FDCD49DcA30c7CF57E578a026d2789',
       factory: '0x9406Cc6185a346906296840746125a0E44976454',
     },
+    bundlerUrl: import.meta.env.VITE_BUNDLER_URL || 'https://rundler-superrelay.fly.dev',
   },
 };